refactor: move name, phone-number, role to users folder
